refactor(e2e): move tests to individual files
