+++ test/e2e/extensionsWithoutCert.test.ts
@@ -3,8 +3,10 @@ import * as path from "path"
 import { getMaybeProxiedCodeServer } from "../utils/helpers"
 import { describe, test, expect } from "./baseFixture"
 
-function runTestExtensionTests() {
-  // This will only work if the test extension is loaded into code-server.
+// Setup
+const flags = ["--extensions-dir", path.join(__dirname, "./extensions")]
+
+describe("Extensions", flags, {}, () => {
   test("should have access to VSCODE_PROXY_URI", async ({ codeServerPage }) => {
     const address = await getMaybeProxiedCodeServer(codeServerPage)
 
@@ -17,22 +19,4 @@ function runTestExtensionTests() {
     const normalizedAddress = address.replace(/\/+$/, "")
     expect(text).toBe(`Info: proxyUri: ${normalizedAddress}/proxy/{{port}}`)
   })
-}
-
-const flags = ["--extensions-dir", path.join(__dirname, "./extensions")]
-
-describe("Extensions", flags, {}, () => {
-  runTestExtensionTests()
 })
-
-if (process.env.USE_PROXY !== "1") {
-  describe("Extensions with --cert", [...flags, "--cert"], {}, () => {
-    runTestExtensionTests()
-  })
-} else {
-  base.describe("Extensions with --cert", () => {
-    base.skip("skipped because USE_PROXY is set", () => {
-      // Playwright will not show this without a function.
-    })
-  })
-}
