+++ test/e2e/githubTokenLoggedOut.test.ts
@@ -2,20 +2,6 @@ import { test as base } from "@playwright/test"
 import { describe, expect, test } from "./baseFixture"
 
 if (process.env.GITHUB_TOKEN) {
-  describe("GitHub token", [], {}, () => {
-    test("should be logged in to pull requests extension", async ({ codeServerPage }) => {
-      await codeServerPage.exec("git init")
-      await codeServerPage.exec("git remote add origin https://github.com/coder/code-server")
-      await codeServerPage.installExtension("GitHub.vscode-pull-request-github")
-      await codeServerPage.executeCommandViaMenus("View: Show Github")
-      await codeServerPage.page.click("text=Sign in")
-      await codeServerPage.page.click("text=Allow")
-      // It should ask to select an account, one of which will be the one we
-      // pre-injected.
-      expect(await codeServerPage.page.isVisible("text=Select an account")).toBe(false)
-    })
-  })
-
   describe("No GitHub token", [], { GITHUB_TOKEN: "" }, () => {
     test("should not be logged in to pull requests extension", async ({ codeServerPage }) => {
       await codeServerPage.exec("git init")
