+++ test/e2e/migrateState.test.ts
@@ -3,7 +3,6 @@ import { promises as fs } from "fs"
 import * as os from "os"
 import * as path from "path"
 import * as util from "util"
-import { getMaybeProxiedCodeServer } from "../utils/helpers"
 import { describe, test, expect } from "./baseFixture"
 import { CodeServer } from "./models/CodeServer"
 
@@ -40,34 +39,6 @@ describe("code-server", [], {}, () => {
 
     return instance
   }
-
-  test("should navigate to home page", async ({ codeServerPage }) => {
-    // We navigate codeServer before each test
-    // and we start the test with a storage state
-    // which means we should be logged in
-    // so it should be on the address
-    const url = codeServerPage.page.url()
-    // We use match because there may be a / at the end
-    // so we don't want it to fail if we expect http://localhost:8080 to match http://localhost:8080/
-    const address = await getMaybeProxiedCodeServer(codeServerPage)
-    expect(url).toMatch(address)
-  })
-
-  test("should always see the code-server editor", async ({ codeServerPage }) => {
-    expect(await codeServerPage.isEditorVisible()).toBe(true)
-  })
-
-  test("should show the Integrated Terminal", async ({ codeServerPage }) => {
-    await codeServerPage.focusTerminal()
-    expect(await codeServerPage.page.isVisible("#terminal")).toBe(true)
-  })
-
-  test("should open a file", async ({ codeServerPage }) => {
-    const dir = await codeServerPage.workspaceDir
-    const file = path.join(dir, "foo")
-    await fs.writeFile(file, "bar")
-    await codeServerPage.openFile(file)
-  })
 
   test("should migrate state to avoid collisions", async ({ codeServerPage }) => {
     // This can take a very long time in development because of how long pages
