+++ test/e2e/terminalAccessVscodeProxyURI.test.ts
@@ -1,5 +1,4 @@
 import * as cp from "child_process"
-import { promises as fs } from "fs"
 import * as path from "path"
 import util from "util"
 import { clean, getMaybeProxiedCodeServer, tmpdir } from "../utils/helpers"
@@ -28,19 +27,5 @@ describe("Integrated Terminal", [], {}, () => {
     const { stdout } = await output
     const address = await getMaybeProxiedCodeServer(codeServerPage)
     expect(stdout).toMatch(address)
-  })
-
-  test("should be able to invoke `code-server` to open a file", async ({ codeServerPage }) => {
-    const tmpFolderPath = await tmpdir(testName)
-    const tmpFile = path.join(tmpFolderPath, "test-file")
-    await fs.writeFile(tmpFile, "test")
-    const fileName = path.basename(tmpFile)
-
-    await codeServerPage.focusTerminal()
-
-    await codeServerPage.page.keyboard.type(`code-server ${tmpFile}`)
-    await codeServerPage.page.keyboard.press("Enter")
-
-    await codeServerPage.waitForTab(fileName)
   })
 })
